MXS-4691: Use 64-bit integers for file positions
Using an int for a file position will result in a narrowing conversion
which in turn causes corrupt binlogs when the file offsets are
misinterpreted either as negative values or as bad offsets into the
file. Using a signed 64-bit integers solves this.

+++ server/modules/routing/pinloki/file_writer.hh
@@ -44,7 +44,7 @@ private:
     {
         std::string   name;
         std::ofstream file;
-        int           write_pos;
+        int64_t       write_pos;
     };
